Pass date to byDateAndActive rather than the queue in order to make the function more reusable.
